requirements.txt
questionary
argparse
crew
pyperclip
threading
itertools